# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "CustomerOrdersController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/me/orders": {
      "get": {
        "operationId": "CustomerOrdersController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sent",
                "accepted",
                "preparing",
                "out_for_delivery",
                "delivered",
                "declined",
                "cancelled"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "in_store"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/me/orders/{id}": {
      "get": {
        "operationId": "CustomerOrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/me/orders/{id}/cancel": {
      "post": {
        "operationId": "CustomerOrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/me/orders/{id}/return": {
      "post": {
        "operationId": "CustomerOrdersController_requestReturn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/me/orders/{id}/rating": {
      "post": {
        "operationId": "CustomerOrdersController_rate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/merchant/orders": {
      "get": {
        "operationId": "MerchantOrdersController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sent",
                "accepted",
                "preparing",
                "out_for_delivery",
                "delivered",
                "declined",
                "cancelled"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "in_store"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/orders/{id}": {
      "get": {
        "operationId": "MerchantOrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/orders/{id}/customer": {
      "get": {
        "operationId": "MerchantOrdersController_customer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/orders/{id}/status": {
      "patch": {
        "operationId": "MerchantOrdersController_status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/orders/{id}/credit-decision": {
      "post": {
        "operationId": "MerchantOrdersController_creditDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/orders/{id}/return-decision": {
      "post": {
        "operationId": "MerchantOrdersController_returnDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/orders/{id}/customer-rating": {
      "post": {
        "operationId": "MerchantOrdersController_rateCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wasl API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "store_id": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "type": "object",
              "properties": {
                "product_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "qty": {
                  "type": "integer",
                  "minimum": 1,
                  "exclusiveMinimum": false,
                  "maximum": 999,
                  "exclusiveMaximum": false
                },
                "note": {
                  "type": "string",
                  "maxLength": 200
                }
              },
              "required": [
                "product_id",
                "qty"
              ]
            }
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "cash",
              "credit",
              "wallet",
              "jeeb",
              "jawali",
              "hasab",
              "onecash"
            ]
          },
          "location": {
            "type": "object",
            "properties": {
              "landmark": {
                "type": "string",
                "minLength": 1,
                "maxLength": 300
              },
              "phone": {
                "type": "string",
                "maxLength": 20
              },
              "label": {
                "type": "string",
                "maxLength": 80
              },
              "lat": {
                "type": "number",
                "minimum": -90,
                "exclusiveMinimum": false,
                "maximum": 90,
                "exclusiveMaximum": false,
                "nullable": true
              },
              "lng": {
                "type": "number",
                "minimum": -180,
                "exclusiveMinimum": false,
                "maximum": 180,
                "exclusiveMaximum": false,
                "nullable": true
              }
            },
            "required": [
              "landmark"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 500
          },
          "wallet_ref": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "store_id",
          "items",
          "payment_method",
          "location"
        ]
      },
      "ReturnDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
      "RatingDto": {
        "type": "object",
        "properties": {
          "stars": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "maximum": 5,
            "exclusiveMaximum": false
          },
          "comment": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "stars"
        ]
      },
      "StatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "preparing",
              "out_for_delivery",
              "delivered",
              "declined"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "DecisionDto": {
        "type": "object",
        "properties": {
          "approve": {
            "type": "boolean"
          }
        },
        "required": [
          "approve"
        ]
      },
...
    }
  }
}